refactor: move cropping functionality in own folder
