+++ ftw/simplelayout/contenttypes/profiles/default/types/ftw.simplelayout.TextBlock.xml
@@ -28,7 +28,7 @@
         <element value="ftw.simplelayout.interfaces.ISimplelayoutBlock" />
         <element value="ftw.simplelayout.contenttypes.behaviors.ITeaser" />
         <element value="plone.app.referenceablebehavior.referenceable.IReferenceable"/>
-        <element value="ftw.simplelayout.images.cropping.IImageCropping"/>
+        <element value="ftw.simplelayout.images.cropping.behaviors.IImageCropping"/>
 
     </property>
