+++ ftw/simplelayout/contenttypes/upgrades/20180822172930_add_image_cropping_behavior_to_text_block/types/ftw.simplelayout.TextBlock.xml
@@ -6,7 +6,7 @@
 
     <!-- enabled behaviors -->
     <property name="behaviors" purge="False">
-        <element value="ftw.simplelayout.images.cropping.IImageCropping"/>
+        <element value="ftw.simplelayout.images.cropping.behaviors.IImageCropping"/>
     </property>
 
 </object>
