+++ ftw/simplelayout/images/cropping/browser/templates/cropping.pt
