+++ ftw/simplelayout/images/cropping/browser/templates/cropping_hard_limit_message.pt
