+++ ftw/simplelayout/images/cropping/browser/templates/cropping_soft_limit_message.pt
